requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gradio==4.44.1

# Data Models & Validation
pydantic==2.6.0
pydantic-settings==2.1.0

# Async
tenacity==8.2.3

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# LLM Integration
litellm==1.18.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Image Processing
Pillow==10.2.0

# Markdown Parsing
markdown==3.5.2

# Utilities
python-multipart==0.0.6
watchdog==3.0.0

# Testing (可选)
pytest==7.4.4
pytest-asyncio==0.23.3

# Development
rich==13.7.0
